backend/requirements.txt
# FarmAura Auto Detect Backend Requirements
# Python 3.8+

# Web Framework
Flask==3.0.0
flask-cors==4.0.0

# Data Processing
pandas==2.1.4
numpy==1.24.3

# Machine Learning
catboost==1.2.2
scikit-learn==1.3.2

# Google Gemini AI
google-generativeai==0.3.2

# Environment & Configuration
python-dotenv==1.0.0

# Utilities
requests==2.31.0

# Database
azure-cosmos==4.5.1
tensorflow==2.15.0
Pillow==10.2.0
